+++ src/runtime/composables/useFormEvents.ts
@@ -1,6 +1,19 @@
 import { inject } from 'vue'
 import { UseEventBusReturn } from '@vueuse/core'
-import { FormEvent } from '../types'
+
+export interface FormError {
+  path: string
+  message: string
+}
+
+export interface Form<T> {
+  validate(): Promise<T>
+}
+
+export interface FormEvent {
+  type: 'blur' // | 'change' | 'focus'
+  path: string
+}
 
 export const useFormEvents = () => {
     const formBus = inject<UseEventBusReturn<FormEvent, string> | undefined>('form-events', undefined)
